增加了库存选项和修改了bug，优化了搜索框

+++ flowers/flowers_transcaction/models.py
@@ -7,7 +7,7 @@ from flowers_base.models import flower_class,flower_data,admin_data
 # 库存信息
 # 出库
 class outbound(models.Model):
-    outbound_id = models.IntegerField(unique=True)
+    outbound_id = models.BigIntegerField(unique=True)
     flowers = models.ForeignKey(flower_data,to_field="flower_name", on_delete=models.CASCADE)
     outbound_num = models.IntegerField()
     outbound_date = models.DateTimeField(auto_now_add=True)
@@ -19,7 +19,7 @@ class outbound(models.Model):
 
 # 入库
 class inbound(models.Model):
-    inbound_id = models.IntegerField(unique=True)
+    inbound_id = models.BigIntegerField(unique=True)
     flowers = models.ForeignKey(flower_data,to_field="flower_name", on_delete=models.CASCADE)
     inbound_num = models.IntegerField()
     inbound_date = models.DateTimeField(auto_now_add=True)
